@@ -4,12 +4,11 @@
   "private": true,
   "dependencies": {
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.1",
-    "styled-components": "^5.3.11",
-    "web-vitals": "^2.1.4"
+    "styled-components": "^6.0.8"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,11 +35,11 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.43.0",
+    "eslint": "^8.50.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0"
   }
 }
