@@ -12,12 +12,12 @@
   "dependencies": {
     "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
-    "pg": "^8.11.0"
+    "pg": "^8.11.3"
   },
   "devDependencies": {
-    "eslint": "^8.41.0",
+    "eslint": "^8.50.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.27.5",
-    "nodemon": "^2.0.22"
+    "eslint-plugin-import": "^2.28.1",
+    "nodemon": "^3.0.1"
   }
 }
