@@ -23,5 +23,5 @@ export const Overlay = styled.div`
 
   animation: ${fadeIn} 0.3s;
 
-  ${({ isLeaving }) => isLeaving && css`animation: ${fadeOut} 0.2s;`}
+  ${({ isLeaving }) => isLeaving && css`animation: ${fadeOut} 0.2s forwards;`}
 `;
