@@ -33,7 +33,7 @@ export const OverLay = styled.div`
   justify-content: center;
   animation: ${fadeIn} 0.3s;
 
-  ${({ isLeaving }) => isLeaving && css` animation ${fadeOut} 0.2s;`}
+  ${({ isLeaving }) => isLeaving && css`animation: ${fadeOut} 0.2s forwards;`}
 `;
 
 export const Container = styled.div`
@@ -44,7 +44,7 @@ export const Container = styled.div`
   background: #FFF;
   box-shadow: 0px 4px 10px rgba(0, 0, 0 ,0.04);
   animation: ${scaleIn} 0.3s;
-  ${({ isLeaving }) => isLeaving && css` animation ${scaleOut} 0.2s;`}
+  ${({ isLeaving }) => isLeaving && css`animation: ${scaleOut} 0.2s forwards;`}
 
   > h1 {
     font-size: 22px;
