@@ -1,12 +1,13 @@
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
+import { memo } from 'react';
 import { Card, ListHeader } from './styles';
 
 import arrow from '../../../../assets/images/icons/arrow.svg';
 import edit from '../../../../assets/images/icons/edit.svg';
 import trash from '../../../../assets/images/icons/trash.svg';
 
-export default function ContactList({
+function ContactList({
   onDeleteContact,
   filteredContacts,
   orderBy,
@@ -69,3 +70,5 @@ ContactList.propTypes = {
   orderBy: PropTypes.string.isRequired,
   onToggleOrderBy: PropTypes.func.isRequired,
 };
+
+export default memo(ContactList);
